pyproject.toml
[project]
name = "portfolio-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "drf-spectacular"
]
